refactor(#86): post 성공,예외 코드 변경
